Add files via upload
The current US Census Bureau world population estimate in June 2019 shows that the current global population is 7,577,130,400 people on earth, which far exceeds the world population of 7.2 billion in 2015. Our own estimate based on UN data shows the world's population surpassing 7.7 billion. 

China is the most populous country in the world with a population exceeding 1.4 billion. It is one of just two countries with a population of more than 1 billion, with India being the second. As of 2018, India has a population of over 1.355 billion people, and its population growth is expected to continue through at least 2050. By the year 2030, the country of India is expected to become the most populous country in the world. This is because India’s population will grow, while China is projected to see a loss in population.

The following 11 countries that are the most populous in the world each have populations exceeding 100 million. These include the United States, Indonesia, Brazil, Pakistan, Nigeria, Bangladesh, Russia, Mexico, Japan, Ethiopia, and the Philippines. Of these nations, all are expected to continue to grow except Russia and Japan, which will see their populations drop by 2030 before falling again significantly by 2050.

Many other nations have populations of at least one million, while there are also countries that have just thousands. The smallest population in the world can be found in Vatican City, where only 801 people reside.

In 2018, the world’s population growth rate was 1.12%. Every five years since the 1970s, the population growth rate has continued to fall. The world’s population is expected to continue to grow larger but at a much slower pace. By 2030, the population will exceed 8 billion. In 2040, this number will grow to more than 9 billion. In 2055, the number will rise to over 10 billion, and another billion people won’t be added until near the end of the century. The current annual population growth estimates from the United Nations are in the millions- estimating that over 80 million new lives are added each year.

This population growth will be significantly impacted by nine specific countries which are situated to contribute to the population growing more quickly than other nations. These nations include the Democratic Republic of the Congo, Ethiopia, India, Indonesia, Nigeria, Pakistan, Uganda, the United Republic of Tanzania, and the United States of America. Particularly of interest, India is on track to overtake China's position as the most populous country by 2030. Additionally, multiple nations within Africa are expected to double their populations before fertility rates begin to slow entirely.

Content

In this Dataset, we have Historical Population data for every Country/Territory in the world by different parameters like Area Size of the Country/Territory, Name of the Continent, Name of the Capital, Density, Population Growth Rate, Ranking based on Population, World Population Percentage, etc.
 
Dataset Glossary (Column-Wise)

 ● Rank: Rank by Population.
 ● CCA3: 3 Digit Country/Territories Code.
 ● Country/Territories: Name of the Country/Territories.
 ● Capital: Name of the Capital.
 ● Continent: Name of the Continent.
 ● 2022 Population: Population of the Country/Territories in the year 2022.
 ● 2020 Population: Population of the Country/Territories in the year 2020.
 ● 2015 Population: Population of the Country/Territories in the year 2015.
 ● 2010 Population: Population of the Country/Territories in the year 2010.
 ● 2000 Population: Population of the Country/Territories in the year 2000.
 ● 1990 Population: Population of the Country/Territories in the year 1990.
 ● 1980 Population: Population of the Country/Territories in the year 1980.
 ● 1970 Population: Population of the Country/Territories in the year 1970.
 ● Area(km²): Area size of the Country/Territories in square kilometer.
 ● Density (per km²): Population Density per square kilometer.
 ● Growth Rate: Population Growth Rate by Country/Territories.
 ● World Population Percentage: The population percentage by each Country/Territories

Acknowledgement

 This Dataset is created from https://worldpopulationreview.com/. If you want to learn more, you can visit the
 Website.
 Cover Photo by: People symbol vector created by rawpixel.com- www.freepik.com